#1 Fixed TypeError by changing int8 to float32 Added information about virtualenv usage Added an empty models folder and requirements.txt

+++ sigmorphon-conv.py
@@ -53,7 +53,7 @@ class MorphonModel:
 
         input_str = Input(shape=(self.max_length,), dtype='int32')
         input_feats = Input(shape=(self.data.feature_vector_length,),
-                            dtype='int8')
+                            dtype='float32')
 
         inputs = [input_str, input_feats]
